update utils component, checkEqual renamed as isDeepEqual and extracted from utils, update DeepComparison component,  chapter 4 tests added
