+++ components/exercises/chapter-4-oscar/DeepComparison/DeepComparison.tsx
@@ -1,24 +1,81 @@
 import React, { useState } from 'react';
-import { checkEqual, lastWeekList, currentWeekList, WeekItem } from './Utils';
+import { isDeepEqual } from './isDeepEqual';
+
+/**
+ * Represents the properties of a week item.
+ */
+export type ShopItem = {
+  name: string;
+  description: string;
+  price: number;
+};
 
 /**
  * This represents an array of arrays of WeekItem objects, essentially creating a two-dimensional array of WeekItem objects. Each element of the outer array is an array of WeekItem objects.
  */
-type GroceryList = WeekItem[][];
+type GroceryList = ShopItem[][];
 
-const newElement: WeekItem = {
+const newElement: ShopItem = {
   name: '🍅 Tomatoes',
   description: '1kg Fresh tomatoes',
   price: 0.5,
 };
 
+export const lastWeekShoppingList: ShopItem[] = [
+  {
+    name: '🥚 Eggs',
+    description: '6x Large class A free range eggs',
+    price: 2.5,
+  },
+  {
+    name: '🐓 Chicken',
+    description: '650gr Fresh Class A skinless chicken breast fillet portions.',
+    price: 4.7,
+  },
+  {
+    name: '🍚 Basmati rice',
+    description:
+      '1kg Aromatic & fluffy A long, slender grain grown in the foothills of the Himalayas',
+    price: 2.5,
+  },
+  {
+    name: '🥔 Potatoes',
+    description: '1kg White potatoes',
+    price: 1.5,
+  },
+];
+
+export const currentWeekShoppingList: ShopItem[] = [
+  {
+    name: '🥚 Eggs',
+    description: '6x Large class A free range eggs',
+    price: 2.5,
+  },
+  {
+    name: '🐓 Chicken',
+    description: '650gr Fresh Class A skinless chicken breast fillet portions.',
+    price: 4.7,
+  },
+  {
+    name: '🍚 Basmati rice',
+    description:
+      '1kg Aromatic & fluffy A long, slender grain grown in the foothills of the Himalayas',
+    price: 2.5,
+  },
+  {
+    name: '🥔 Potatoes',
+    description: '1kg White potatoes',
+    price: 1.5,
+  },
+];
+
 /**
- * Renders a component that compares two grocery lists and allows adding/removing items.
+ * Renders a component that compares two grocery lists and allows adding/removing items and checks for deep equality of both grocery lists.
  */
 export default function DeepComparison(): JSX.Element {
   const [groceryList, setGroceryList] = useState<GroceryList>([
-    lastWeekList,
-    currentWeekList,
+    lastWeekShoppingList,
+    currentWeekShoppingList,
   ]);
 
   /**
@@ -51,19 +108,19 @@ export default function DeepComparison(): JSX.Element {
 
           {groceryList &&
             groceryList[0].map((item) => (
-              <ol key={item.name}>
+              <ul key={item.name}>
                 <li className='relative mt-[24px] rounded-lg bg-yellow-200 p-2 pl-[16px] font-[16px] leading-[20px] text-black'>
                   <strong>{item.name}</strong>
                   <p className='mt-2 text-sm'>{item.description}</p>
                   <p className='mt-1 text-right'>{item.price} $</p>
                 </li>
-              </ol>
+              </ul>
             ))}
         </div>
 
         <div className='mx-[10px] my-auto flex w-[300px] flex-1 flex-col rounded-md border-solid p-[10px] shadow-lg'>
           <div className='flex flex-row justify-between'>
-            <strong>Grocery of last week</strong>
+            <strong>Grocery of this week</strong>
             <div>
               <button
                 className='rounded-s-full bg-green-300 p-1'
@@ -82,26 +139,24 @@ export default function DeepComparison(): JSX.Element {
 
           {groceryList &&
             groceryList[1]?.map((item) => (
-              <ol key={item.name}>
+              <ul key={item.name}>
                 <li className='relative mt-[24px] rounded-lg bg-blue-200 p-2 pl-[16px] font-[16px] leading-[20px] text-black'>
                   <strong>{item.name}</strong>
                   <p className='mt-2 text-sm'>{item.description}</p>
                   <p className='mt-1 text-right'>{item.price} $</p>
                 </li>
-              </ol>
+              </ul>
             ))}
         </div>
       </div>
 
-      {checkEqual(groceryList[0], groceryList[1]) === true ? (
-        <span className='mt-4 cursor-pointer rounded-lg bg-green-300 p-2 shadow-md'>
+      {isDeepEqual(groceryList[0], groceryList[1]) === true ? (
+        <span className='mt-4 rounded-lg bg-green-300 p-2 shadow-md'>
           It is the same list
-          {checkEqual(lastWeekList, currentWeekList)}
         </span>
       ) : (
-        <span className='cursor-pointer rounded-lg bg-red-300 p-2 shadow-md '>
+        <span className='mt-4 rounded-lg bg-red-300 p-2 shadow-md '>
           It is not the same list
-          {checkEqual(groceryList[1], groceryList[2])}
         </span>
       )}
     </div>
