+++ components/exercises/chapter-4-oscar/Utils.test.tsx
@@ -10,7 +10,6 @@ import {
   ListItemProp,
   getFood,
   ListItem,
-  checkEqual,
 } from './Utils';
 import { render, screen } from '@testing-library/react';
 
@@ -274,47 +273,5 @@ describe('getFood', () => {
     render(getFood(items));
     const foodElement = screen.getByText('Goulash');
     expect(foodElement).toBeInTheDocument();
-  });
-});
-
-describe('checkEqual', () => {
-  it('should return true for equal values', () => {
-    const obj1 = { name: 'John', age: 30 };
-    const obj2 = { name: 'John', age: 30 };
-    expect(checkEqual(obj1, obj2)).toBe(true);
-  });
-
-  it('should return false for different values', () => {
-    const obj1 = { name: 'John', age: 30 };
-    const obj2 = { name: 'John', age: 25 };
-    expect(checkEqual(obj1, obj2)).toBe(false);
-  });
-
-  it('should return true for deeply equal values', () => {
-    const obj1 = {
-      name: 'John',
-      age: 30,
-      address: { city: 'New York', country: 'USA' },
-    };
-    const obj2 = {
-      name: 'John',
-      age: 30,
-      address: { city: 'New York', country: 'USA' },
-    };
-    expect(checkEqual(obj1, obj2)).toBe(true);
-  });
-
-  it('should return false for deeply different values', () => {
-    const obj1 = {
-      name: 'John',
-      age: 30,
-      address: { city: 'New York', country: 'USA' },
-    };
-    const obj2 = {
-      name: 'John',
-      age: 30,
-      address: { city: 'Los Angeles', country: 'USA' },
-    };
-    expect(checkEqual(obj1, obj2)).toBe(false);
   });
 });
